+++ src/main/java/io/ikatoo/models/dao/ProfessorDAO.java
@@ -5,13 +5,12 @@ package io.ikatoo.models.dao;
  */
 
 import io.ikatoo.models.Professor;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
 
 @Component
-public interface ProfessorDAO extends PagingAndSortingRepository<Professor, Integer> {
+public interface ProfessorDAO extends PagingAndSortingRepository<Professor, Long> {
     List<Professor> findByProfessorIgnoreCaseContaining(String professor);
 }
